pyproject.toml
[tool.poetry]
name = "cell-kn-mvp-etl-ontologies"
version = "0.1.0"
description = "Load ontologies into an ArangoDB instance"
authors = ["Raymond LeClair <raymond.leclair@springbok.io>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
arango-rdf = ">=1.0"
cellxgene-census = ">=1.14"
gget = ">=0.29"
ontogpt = ">=0.3"
plotly = ">=5.22"
pybiomart = ">=0.2"
python-arango = ">=7.4,<8.0.0"
rdflib = ">=7.0"
scanpy = ">=1.10"
scikit-misc = ">=0.3"

[tool.poetry.group.dev.dependencies]
black = ">=24.4"
flake8 = ">=7.0"
ipython = ">=8.25"
jupyterlab = ">=4.2"
mypy = ">=1.10"
notebook = ">=7.2"
pylint = ">=3.2"
pytest = ">=8.2"
sphinx = ">=8.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
